Update питання 1.py

+++ питання 1.py
@@ -1,4 +1,4 @@
-a = int(input('Length in centimeters:'))
+a = float(input('Length in centimeters:'))
 if a < 0:
     print('The entry is invalid')
 if a > 0:
